requirements.txt
python-telegram-bot[all]
python-telegram-bot[ext]
python-dotenv
Flask
psycopg2